Make STP work for both NetworkService enabled/disabled codepaths.
