+++ browser/net/brave_network_delegate_base.cc
@@ -10,6 +10,7 @@
 
 #include "base/task/post_task.h"
 #include "brave/browser/brave_browser_process_impl.h"
+#include "brave/browser/net/brave_stp_util.h"
 #include "brave/common/pref_names.h"
 #include "brave/components/brave_shields/browser/brave_shields_util.h"
 #include "brave/components/brave_shields/browser/brave_shields_web_contents_observer.h"
@@ -22,7 +23,6 @@
 #include "components/prefs/pref_service.h"
 #include "content/public/browser/browser_task_traits.h"
 #include "content/public/browser/browser_thread.h"
-#include "net/base/registry_controlled_domains/registry_controlled_domain.h"
 #include "net/url_request/url_request.h"
 
 using content::BrowserThread;
@@ -66,36 +66,6 @@ bool OnAllowAccessCookies(
 
 }  // namespace
 
-base::flat_set<base::StringPiece>* TrackableSecurityHeaders() {
-  static base::NoDestructor<base::flat_set<base::StringPiece>>
-      kTrackableSecurityHeaders(base::flat_set<base::StringPiece>{
-          "Strict-Transport-Security", "Expect-CT", "Public-Key-Pins",
-          "Public-Key-Pins-Report-Only"});
-  return kTrackableSecurityHeaders.get();
-}
-
-void RemoveTrackableSecurityHeadersForThirdParty(
-    const GURL& request_url, const url::Origin& top_frame_origin,
-    const net::HttpResponseHeaders* original_response_headers,
-    scoped_refptr<net::HttpResponseHeaders>* override_response_headers) {
-  if (!original_response_headers && !override_response_headers->get()) {
-    return;
-  }
-
-  if (net::registry_controlled_domains::SameDomainOrHost(
-          request_url, top_frame_origin,
-          net::registry_controlled_domains::INCLUDE_PRIVATE_REGISTRIES)) {
-    return;
-  }
-
-  if (!override_response_headers->get()) {
-    *override_response_headers =
-        new net::HttpResponseHeaders(original_response_headers->raw_headers());
-  }
-  for (auto header : *TrackableSecurityHeaders()) {
-    (*override_response_headers)->RemoveHeader(header.as_string());
-  }
-}
 
 BraveNetworkDelegateBase::BraveNetworkDelegateBase(
     extensions::EventRouterForwarder* event_router)
@@ -190,7 +160,7 @@ int BraveNetworkDelegateBase::OnHeadersReceived(
     scoped_refptr<net::HttpResponseHeaders>* override_response_headers,
     GURL* allowed_unsafe_redirect_url) {
   if (request->top_frame_origin().has_value()) {
-    RemoveTrackableSecurityHeadersForThirdParty(
+    brave::RemoveTrackableSecurityHeadersForThirdParty(
         request->url(), request->top_frame_origin().value(),
         original_response_headers, override_response_headers);
   }
