+++ browser/net/brave_network_delegate_base.h
@@ -11,9 +11,6 @@
 #include <string>
 #include <vector>
 
-#include "base/containers/flat_set.h"
-#include "base/files/file_path.h"
-#include "base/strings/string_piece.h"
 #include "brave/browser/net/url_context.h"
 #include "chrome/browser/net/chrome_network_delegate.h"
 #include "content/public/browser/browser_thread.h"
@@ -28,13 +25,6 @@ class EventRouterForwarder;
 namespace net {
 class URLRequest;
 }
-
-base::flat_set<base::StringPiece>* TrackableSecurityHeaders();
-
-void RemoveTrackableSecurityHeadersForThirdParty(
-    const GURL& request_url, const url::Origin& top_frame_origin,
-    const net::HttpResponseHeaders* original_response_headers,
-    scoped_refptr<net::HttpResponseHeaders>* override_response_headers);
 
 // BraveNetworkDelegateBase is the central point from within the Brave code to
 // add hooks into the network stack.
