+++ browser/net/brave_network_delegate_base_unittest.cc
@@ -8,12 +8,16 @@
 #include <string>
 
 #include "brave/browser/net/url_context.h"
+#include "brave/browser/net/brave_stp_util.h"
 #include "chrome/test/base/chrome_render_view_host_test_harness.h"
 #include "net/traffic_annotation/network_traffic_annotation_test_helper.h"
 #include "net/url_request/url_request_test_util.h"
 #include "url/gurl.h"
 
+using brave::RemoveTrackableSecurityHeadersForThirdParty;
+using brave::TrackableSecurityHeaders;
 using net::HttpResponseHeaders;
+
 
 namespace {
